% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bi.r.R
\name{init}
\alias{init}
\title{create domo environment and store login info}
\usage{
init(customer, token, config = NULL, verbose = FALSE)
}
\arguments{
\item{customer}{The customer ID or base URL.  e.g.  acme.domo.com  or  acme}

\item{token}{The DEV token required for API access.}
}
\description{
create domo environment and store login info
}
\examples{
init(Sys.getenv('DOMO_BASE_URL'), Sys.getenv('DEVELOPER_TOKEN'))
}
